PartDesign: Preserve standalone thickness placement
Restore placement from the resolved dress-up base so transformed standalone Thickness objects remain aligned when OCC operates in local coordinates. Add regression coverage, consolidate Thickness test discovery, and keep the recto-verso helper private to FeatureThickness.cpp.
