+++ src/Mod/PartDesign/App/CMakeLists.txt
@@ -69,8 +69,6 @@ SET(FeaturesDressUp_SRCS
     FeatureDraft.h
     FeatureThickness.cpp
     FeatureThickness.h
-    RectoVersoThickness.cpp
-    RectoVersoThickness.h
 )
 SOURCE_GROUP("DressUpFeatures" FILES ${FeaturesDressUp_SRCS})
