+++ src/Mod/PartDesign/TestPartDesignApp.py
@@ -54,10 +54,7 @@ from PartDesignTests.TestBaseFeature import TestBaseFeature
 from PartDesignTests.TestFillet import TestFillet
 from PartDesignTests.TestChamfer import TestChamfer
 from PartDesignTests.TestDraft import TestDraft
-from PartDesignTests.TestThickness import (
-    TestRectoVersoThickness,
-    TestThickness,
-)
+from PartDesignTests.TestThickness import TestThickness
 
 # extras
 from PartDesignTests.TestInvoluteGear import TestInvoluteGear
